Consolidate the two SSHTunnelManager classes into one in mngr_forward
The reverse-tunnel SSHTunnelManager in libs/mngr_latchkey was a near-
verbatim copy of the one in libs/mngr_forward; the two diverged on
three things only -- exponential backoff for the repair loop, an
agent_id tag on each ReverseTunnelInfo, and a
remove_reverse_tunnels_for_agent method used by the latchkey
destruction path. All three are strictly nicer behavior so they've
moved into the mngr_forward manager, which is now the single SSH
tunneling implementation in the monorepo.

Changes:
- libs/mngr_forward/imbue/mngr_forward/ssh_tunnel.py:
  - ReverseTunnelInfo gains an optional agent_id: str | None
    field.
  - setup_reverse_tunnel gains an optional agent_id parameter
    (default None preserves existing call-site behavior).
  - The repair loop uses per-tunnel exponential backoff capped at 5
    minutes instead of a flat 30s cadence. Successful repair clears
    the backoff bookkeeping; an alive-sibling observation on a
    previously-failed tunnel also clears stale state so the next
    break starts a fresh schedule.
  - New remove_reverse_tunnels_for_agent(agent_id) tears down all
    reverse tunnels tagged with a given agent_id. It checks the
    forward-tunnel socket map before closing the underlying SSH
    client so a live forward tunnel on the same host is not
    disturbed.
- libs/mngr_latchkey/imbue/mngr_latchkey/ssh_tunnel.py: deleted.
  Its 635-line test file is also deleted; the latchkey-specific
  tests (exponential backoff, remove_reverse_tunnels_for_agent,
  alive-sibling failure-state clearing) are now part of
  libs/mngr_forward/imbue/mngr_forward/ssh_tunnel_test.py.
- libs/mngr_latchkey/imbue/mngr_latchkey/{discovery,cli,
  discovery_stream,discovery_stream_test,core_test}.py: switch
  RemoteSSHInfo / SSHTunnelError / SSHTunnelManager imports from
  imbue.mngr_latchkey.ssh_tunnel to imbue.mngr_forward.ssh_tunnel.
- libs/mngr_forward/imbue/mngr_forward/reverse_handler_test.py: the
  test-double SSHTunnelManager grows the matching agent_id kwarg on
  its setup_reverse_tunnel override.
- apps/minds/imbue/minds/desktop_client/ssh_tunnel.py: docstring
  updated to point at the single tunnel manager.
- libs/mngr_latchkey/README.md: import example updated.

Tests:
  PYTEST_MAX_DURATION_SECONDS=300 uv run pytest -n 4 --dist=worksteal \
    --max-worker-restart=0 -m 'not acceptance and not release and not \
    minds_deployment and not minds_services' --no-cov \
    apps/minds/ libs/mngr_latchkey/ libs/mngr_forward/ libs/mngr_imbue_cloud/
  1417 passed, 2 failed (the two failures are pre-existing -- the
  pool_test.py tests require ssh-keygen on PATH, which is not
  installed in this sandbox; the failing files are unchanged by this
  diff).
